refactor(space): Move `pdist` into own function
